[FEATURE] Add ability to specify HTML fetch URL to avoid auto-resolving
Allows a URL to be specified as source for the HTML even when the
"live identifier" is set to a local page. The extension reads the HTML
from that exact URL instead of attempting to automatically build the
URL based on Site configuration.

Sometimes necessary if site is running in a container or otherwise is
behind a reverse proxy - for example, when running in DDEV setups.